refactor: move FilterSliderInput component to Filter folder
